feat: include `security` in output when using `inverse` and `valid` options (#15)
* feat: include `security` in output when using `inverse` and `valid` options

Using the `inverse` and `valid` options together generally works well,
but excludes the `security` property from the filtered spec.  Since
`security` is an array rather than an object, it isn't possible add the
filter tag to it in the source spec to get it included. This change
implements logic to copy relevant security requirements (i.e. those
whose scheme/definition is present in the filtered spec) from the source
document.

* fix: ensure `components` exists before trying to look inside it